requirements.txt
agency-swarm[fastapi]>=1.4.0
agency-swarm[litellm]>=1.4.0
youtube-transcript-api
python-dotenv
notion-client
